refactor: rename crypto.h as peer-mse.h
